Move the PostHog bootstrapping stuff into its own wrapper component
